Resolved conflicts. (28.01.24)
