--- src/test/java/school/faang/user_service/service/UserServiceTest.java
@@ -1,13 +1,0 @@
-package school.faang.user_service.service;
-
-import org.junit.jupiter.api.Test;
-
-import static org.junit.jupiter.api.Assertions.*;
-
-class UserServiceTest {
-
-    @Test
-    void getUsers() {
-        assertTrue(true);
-    }
-}
