+++ src/test/java/school/faang/user_service/service/event/EventServiceTest.java
@@ -1,6 +1,7 @@
 package school.faang.user_service.service.event;
 
 import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -11,14 +12,20 @@ import org.mockito.junit.jupiter.MockitoExtension;
 import org.mockito.junit.jupiter.MockitoSettings;
 import org.mockito.quality.Strictness;
 import school.faang.user_service.dto.event.EventDto;
+import school.faang.user_service.dto.event.EventFilterDto;
 import school.faang.user_service.entity.Skill;
 import school.faang.user_service.entity.User;
 import school.faang.user_service.entity.event.Event;
 import school.faang.user_service.exception.DataValidationException;
+import school.faang.user_service.filter.event.EventFilter;
+import school.faang.user_service.filter.event.EventOwnerIdPattern;
+import school.faang.user_service.filter.event.EventStartDatePattern;
+import school.faang.user_service.filter.event.EventTitlePattern;
 import school.faang.user_service.mapper.event.EventMapper;
 import school.faang.user_service.repository.event.EventRepository;
 import school.faang.user_service.validator.event.EventValidator;
 
+import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
@@ -39,6 +46,56 @@ class EventServiceTest {
     private EventValidator eventValidator;
     @InjectMocks
     private EventService eventService;
+
+    private List<EventFilter> eventFilters = new ArrayList<>();
+    private LocalDateTime startDate;
+    private User owner;
+    private Skill skill;
+    private Event eventFirst;
+    private Event eventSecond;
+    private EventFilterDto eventFilterDtoId;
+    private EventFilterDto eventFilterDtoTitle;
+    private EventFilterDto eventFilterDtoStartDate;
+
+    @BeforeEach
+    void setUp() {
+        eventFilters.add(new EventOwnerIdPattern());
+        eventFilters.add(new EventTitlePattern());
+        eventFilters.add(new EventStartDatePattern());
+
+        eventService = new EventService(eventRepository, eventMapper, eventValidator, eventFilters);
+
+        startDate = LocalDateTime.now().plusDays(1L);
+        owner = User.builder().id(1L).active(true).build();
+        skill = Skill.builder().id(1L).build();
+
+        eventFirst = Event.builder()
+                .id(1L)
+                .title("EventFirst")
+                .maxAttendees(2)
+                .owner(owner)
+                .relatedSkills(List.of(skill))
+                .startDate(LocalDateTime.now())
+                .build();
+        eventSecond = Event.builder()
+                .id(1L)
+                .title("EventSecond")
+                .maxAttendees(2)
+                .owner(owner)
+                .relatedSkills(List.of(skill))
+                .startDate(startDate)
+                .build();
+
+        eventFilterDtoId = EventFilterDto.builder()
+                .ownerPattern(1L)
+                .build();
+        eventFilterDtoTitle = EventFilterDto.builder()
+                .titlePattern("EventSecond")
+                .build();
+        eventFilterDtoStartDate = EventFilterDto.builder()
+                .startDatePattern(startDate)
+                .build();
+    }
 
     @Test
     public void successCreateAndSaveEventWhenTwoPassedValidation() {
@@ -193,6 +250,27 @@ class EventServiceTest {
         long wrongId = 15L;
 
         Mockito.verify(eventRepository, Mockito.never()).deleteById(wrongId);
+    }
+
+    @Test
+    public void successGetEvents_ByIdFilter() {
+        Mockito.when(eventRepository.findAll()).thenReturn(List.of(eventFirst, eventSecond));
+        List<EventDto> resultToFilterById = eventService.getEventsByFilter(eventFilterDtoId);
+        assertEquals(2, resultToFilterById.size());
+    }
+
+    @Test
+    public void successGetEvents_ByTitleFilter() {
+        Mockito.when(eventRepository.findAll()).thenReturn(List.of(eventFirst, eventSecond));
+        List<EventDto> resultToFilterByTitle = eventService.getEventsByFilter(eventFilterDtoTitle);
+        assertEquals(1, resultToFilterByTitle.size());
+    }
+
+    @Test
+    public void successGetEvents_ByStartDateFilter() {
+        Mockito.when(eventRepository.findAll()).thenReturn(List.of(eventFirst, eventSecond));
+        List<EventDto> resultToFilterByStartDate = eventService.getEventsByFilter(eventFilterDtoStartDate);
+        assertEquals(1, resultToFilterByStartDate.size());
     }
 
 }
